Moving Options to model package
