+++ sdk/communication/azure-communication-networktraversal/src/samples/java/com/azure/communication/networktraversal/CreateAndIssueRelayCredentialsExample.java
@@ -7,6 +7,7 @@ import com.azure.communication.identity.CommunicationIdentityClient;
 import com.azure.communication.identity.CommunicationIdentityClientBuilder;
 import com.azure.communication.networktraversal.models.CommunicationRelayConfiguration;
 import com.azure.communication.networktraversal.models.CommunicationIceServer;
+import com.azure.communication.networktraversal.models.GetRelayConfigurationOptions;
 import java.util.List;
 
 /**
